Move ResolveFilter and impls to render::filters
